Crate hex_microservice

Hexagonal microservice example

A simple URL shortener ported from Go. Demonstrates some features of Hexagonal architecture (and Clean Architecture, which overlaps pleasantly)

Modules

api
api defines an http adapter for a url shortening service.

repository
repository provides concrete implementations of the RedirectRepository trait

serializer
serializer provides concrete implementations of the RedirectSerializer interface

short_url
short_url defines the domain logic and ports for a url shortening service.