Module hex_microservice::short_url

short_url defines the domain logic and ports for a url shortening service.

Modules

error
Define custom Error types for the short url domain

Structs

Redirect
A URL redirect model

Service
A URL redirect service

Traits

RedirectRepository
Provides a find method for looking up URLs based on their short code and a store method for saving Redirect objects

RedirectSerializer
Provides methods to serialize and deserialize Redirect objects

RedirectService
Provides a find method for looking up URLs based on their short code and a store method for saving Redirect objects