Trait hex_microservice::short_url::RedirectRepository

pub trait RedirectRepository {
    fn find(&self, code: &str) -> Result<Redirect, RedirectErr>;

    fn store(&self, redirect: &Redirect) -> Result<(), RedirectErr>;
}

Provides a find method for looking up URLs based on their short code and a store method for saving Redirect objects

Required methods

fn find(&self, code: &str) -> Result<Redirect, RedirectErr>

Retrieve a URL from the repo based on its short code

fn store(&self, redirect: &Redirect) -> Result<(), RedirectErr>

Save a Redirect object to the repository

Loading content...